+++ internal/js_parser/js_parser_lower_test.go
@@ -3,6 +3,8 @@ package js_parser
 import (
 	"fmt"
 	"testing"
+
+	"github.com/evanw/esbuild/internal/compat"
 )
 
 func TestLowerFunctionArgumentScope(t *testing.T) {
@@ -659,4 +661,48 @@ func TestLowerOptionalCatchBinding(t *testing.T) {
 func TestLowerExportStarAs(t *testing.T) {
 	expectPrintedTarget(t, 2020, "export * as ns from 'path'", "export * as ns from \"path\";\n")
 	expectPrintedTarget(t, 2019, "export * as ns from 'path'", "import * as ns from \"path\";\nexport { ns };\n")
+}
+
+func TestAsyncGeneratorFns(t *testing.T) {
+	err := ""
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait, "async function gen() {}", err)
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait, "(async function () {});", err)
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait, "({ async foo() {} });", err)
+
+	err = "<stdin>: ERROR: Transforming generator functions to the configured target environment is not supported yet\n"
+	expectParseErrorWithUnsupportedFeatures(t, compat.Generator, "function* gen() {}", err)
+	expectParseErrorWithUnsupportedFeatures(t, compat.Generator, "(function* () {});", err)
+	expectParseErrorWithUnsupportedFeatures(t, compat.Generator, "({ *foo() {} });", err)
+
+	err = "<stdin>: ERROR: Transforming async functions to the configured target environment is not supported yet\n"
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait|compat.Generator, "async function gen() {}", err)
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait|compat.Generator, "(async function () {});", err)
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait|compat.Generator, "({ async foo() {} });", err)
+
+	err = "<stdin>: ERROR: Transforming async generator functions to the configured target environment is not supported yet\n"
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncGenerator, "async function* gen() {}", err)
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncGenerator, "(async function* () {});", err)
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncGenerator, "({ async *foo() {} });", err)
+}
+
+func TestForAwait(t *testing.T) {
+	err := ""
+	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait, "async function gen() { for await (x of y) ; }", err)
+	expectParseErrorWithUnsupportedFeatures(t, compat.Generator, "async function gen() { for await (x of y) ; }", err)
+
+	// This is ok because for-await can be lowered to await
+	expectParseErrorWithUnsupportedFeatures(t, compat.ForAwait|compat.Generator, "async function gen() { for await (x of y) ; }", err)
+
+	// This is ok because for-await can be lowered to yield
+	expectParseErrorWithUnsupportedFeatures(t, compat.ForAwait|compat.AsyncAwait, "async function gen() { for await (x of y) ; }", err)
+
+	// This is not ok because for-await can't be lowered
+	err =
+		"<stdin>: ERROR: Transforming async functions to the configured target environment is not supported yet\n" +
+			"<stdin>: ERROR: Transforming for-await loops to the configured target environment is not supported yet\n"
+	expectParseErrorWithUnsupportedFeatures(t, compat.ForAwait|compat.AsyncAwait|compat.Generator, "async function gen() { for await (x of y) ; }", err)
+
+	// Can't use for-await at the top-level without top-level await
+	err = "<stdin>: ERROR: Top-level await is not available in the configured target environment\n"
+	expectParseErrorWithUnsupportedFeatures(t, compat.TopLevelAwait, "for await (x of y) ;", err)
 }