+++ internal/js_parser/js_parser_test.go
@@ -5235,28 +5235,6 @@ func TestES5(t *testing.T) {
 		"<stdin>: ERROR: Transforming generator functions to the configured target environment is not supported yet\n")
 }
 
-func TestAsyncGeneratorFns(t *testing.T) {
-	err := ""
-	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait, "async function gen() {}", err)
-	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait, "(async function () {});", err)
-	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait, "({ async foo() {} });", err)
-
-	err = "<stdin>: ERROR: Transforming generator functions to the configured target environment is not supported yet\n"
-	expectParseErrorWithUnsupportedFeatures(t, compat.Generator, "function* gen() {}", err)
-	expectParseErrorWithUnsupportedFeatures(t, compat.Generator, "(function* () {});", err)
-	expectParseErrorWithUnsupportedFeatures(t, compat.Generator, "({ *foo() {} });", err)
-
-	err = "<stdin>: ERROR: Transforming async functions to the configured target environment is not supported yet\n"
-	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait|compat.Generator, "async function gen() {}", err)
-	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait|compat.Generator, "(async function () {});", err)
-	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncAwait|compat.Generator, "({ async foo() {} });", err)
-
-	err = "<stdin>: ERROR: Transforming async generator functions to the configured target environment is not supported yet\n"
-	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncGenerator, "async function* gen() {}", err)
-	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncGenerator, "(async function* () {});", err)
-	expectParseErrorWithUnsupportedFeatures(t, compat.AsyncGenerator, "({ async *foo() {} });", err)
-}
-
 func TestASCIIOnly(t *testing.T) {
 	es5 := "<stdin>: ERROR: \"𐀀\" cannot be escaped in the configured target environment " +
 		"but you can set the charset to \"utf8\" to allow unescaped Unicode characters\n"